# src/OpenMetrc.Scraper/OpenApi/metrc-swagger-v2.yaml
swagger: 2.0
info:
  title: METRC API
  version: v2
paths:
...
  /plantbatches/v2/:
    delete:
      tags:
      - PlantBatches
      summary: ''
      description: 'Completes the destruction of plant batches based on the provided input data.</br><b>Permissions Required</b>: View Immature Plants • Destroy Immature Plants'
      externalDocs:
        x-doc-key: delete_plantbatches_v2_
      operationId: DeletePlantBatches
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to destroy the Plant Batch.
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/active:
    get:
      tags:
      - PlantBatches
      summary: ''
      description: 'Retrieves a list of active plant batches for the specified Facility, optionally filtered by last modified date.</br><b>Permissions Required</b>: View Immature Plants'
      externalDocs:
        x-doc-key: get_plantbatches_v2_active
      operationId: GetPlantBatchActive
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of active plant batches.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /plantbatches/v2/{id}:
    get:
      tags:
      - PlantBatches
      summary: ''
      description: 'Retrieves a Plant Batch by Id.</br><b>Permissions Required</b>: View Immature Plants'
      externalDocs:
        x-doc-key: get_plantbatches_v2_{id}
      operationId: GetPlantBatchById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        description: If specified, the Plant Batch will be validated against the specified License Number. If not specified, the Plant Batch will be validated against all of the User's current Facilities. Please note that if the Plant Batch is not valid for the specified License Number, a 401 Unauthorized status will be returned.
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/inactive:
    get:
      tags:
      - PlantBatches
      summary: ''
      description: 'Retrieves a list of inactive plant batches for the specified Facility, optionally filtered by last modified date.</br><b>Permissions Required</b>: View Immature Plants'
      externalDocs:
        x-doc-key: get_plantbatches_v2_inactive
      operationId: GetPlantBatchInactive
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of inactive plant batches.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /plantbatches/v2/types:
    get:
      tags:
      - PlantBatches
      summary: ''
      description: 'Retrieves a list of plant batch types.</br><b>Permissions Required</b>: <i>none</i>'
      externalDocs:
        x-doc-key: get_plantbatches_v2_types
      operationId: GetPlantBatchTypes
      parameters:
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/waste:
    get:
      tags:
      - PlantBatches
      summary: ''
      description: 'Retrieves waste details associated with plant batches at a specified Facility.</br><b>Permissions Required</b>: View Plants Waste'
      externalDocs:
        x-doc-key: get_plantbatches_v2_waste
      operationId: GetPlantBatchWaste
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of waste plants.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      responses:
        200:
          description: Ok
    post:
      tags:
      - PlantBatches
      summary: ''
      description: 'Records waste information for plant batches based on the submitted data for the specified Facility.</br><b>Permissions Required</b>: Manage Plants Waste'
      externalDocs:
        x-doc-key: post_plantbatches_v2_waste
      operationId: PostPlantBatchWaste
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to record waste of the Plant Batch.
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/waste/reasons:
    get:
      tags:
      - PlantBatches
      summary: ''
      description: 'Retrieves a list of valid waste reasons associated with immature plant batches for the specified Facility.</br><b>Permissions Required</b>: <i>none</i>'
      externalDocs:
        x-doc-key: get_plantbatches_v2_waste_reasons
      operationId: GetPlantBatchWasteReasons
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return a list of waste reasons.
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/additives:
    post:
      tags:
      - PlantBatches
      summary: ''
      description: 'Records Additive usage details for plant batches at a specific Facility.</br><b>Permissions Required</b>: Manage Plants Additives'
      externalDocs:
        x-doc-key: post_plantbatches_v2_additives
      operationId: PostPlantBatchAdditives
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to record plant additives.
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/additives/usingtemplate:
    post:
      tags:
      - PlantBatches
      summary: ''
      description: 'Records Additive usage for plant batches at a Facility using predefined additive templates.</br><b>Permissions Required</b>: Manage Plants Additives'
      externalDocs:
        x-doc-key: post_plantbatches_v2_additives_usingtemplate
      operationId: PostPlantBatchAdditiveUsingtemplate
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to record plant additives.
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/adjust:
    post:
      tags:
      - PlantBatches
      summary: ''
      description: 'Applies Facility specific adjustments to plant batches based on submitted reasons and input data.</br><b>Permissions Required</b>: View Immature Plants • Manage Immature Plants Inventory'
      externalDocs:
        x-doc-key: post_plantbatches_v2_adjust
      operationId: PostPlantBatchAdjust
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to record adjustments.
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/growthphase:
    post:
      tags:
      - PlantBatches
      summary: ''
      description: 'Updates the growth phase of plants at a specified Facility based on tracking information.</br><b>Permissions Required</b>: View Immature Plants • Manage Immature Plants Inventory • View Veg/Flower Plants • Manage Veg/Flower Plants Inventory'
      externalDocs:
        x-doc-key: post_plantbatches_v2_growthphase
      operationId: PostPlantBatchGrowthPhase
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to change the growth phase.
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/packages:
    post:
      tags:
      - PlantBatches
      summary: ''
      description: 'Creates packages from plant batches at a Facility, with optional support for packaging from mother plants.</br><b>Permissions Required</b>: View Immature Plants • Manage Immature Plants Inventory • View Packages • Create/Submit/Discontinue Packages'
      externalDocs:
        x-doc-key: post_plantbatches_v2_packages
      operationId: PostPlantBatchPackages
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to create a package from the Plant Batch.
        format: ''
      - type: boolean
        name: isFromMotherPlant
        in: query
        description: Whether the Package contains a Plant Batch from a mother plant - Boolean. Reference v2/packages/frommotherplant
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/packages/frommotherplant:
    post:
      tags:
      - PlantBatches
      summary: ''
      description: 'Creates packages from mother plants at the specified Facility.</br><b>Permissions Required</b>: View Immature Plants • Manage Immature Plants Inventory • View Packages • Create/Submit/Discontinue Packages'
      externalDocs:
        x-doc-key: post_plantbatches_v2_packages_frommotherplant
      operationId: PostPlantBatchPackageFromMotherPlant
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to create packages from a mother plant.
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/plantings:
    post:
      tags:
      - PlantBatches
      summary: ''
      description: 'Creates new plantings for a Facility by generating plant batches based on provided planting details.</br><b>Permissions Required</b>: View Immature Plants • Manage Immature Plants Inventory'
      externalDocs:
        x-doc-key: post_plantbatches_v2_plantings
      operationId: PostPlantBatchPlantings
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to create the plantings.
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/split:
    post:
      tags:
      - PlantBatches
      summary: ''
      description: 'Splits an existing Plant Batch into multiple groups at the specified Facility.</br><b>Permissions Required</b>: View Immature Plants • Manage Immature Plants Inventory'
      externalDocs:
        x-doc-key: post_plantbatches_v2_split
      operationId: PostPlantBatchSplit
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to split the plant batches.
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/location:
    put:
      tags:
      - PlantBatches
      summary: ''
      description: 'Moves one or more plant batches to new locations with in a specified Facility.</br><b>Permissions Required</b>: View Immature Plants • Manage Immature Plants'
      externalDocs:
        x-doc-key: put_plantbatches_v2_location
      operationId: PutPlantBatchLocation
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to change the Location of the Plant Batch.
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/name:
    put:
      tags:
      - PlantBatches
      summary: ''
      description: 'Renames plant batches at a specified Facility.</br><b>Permissions Required</b>: View Veg/Flower Plants • Manage Veg/Flower Plants Inventory'
      externalDocs:
        x-doc-key: put_plantbatches_v2_name
      operationId: PutPlantBatchName
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to rename the Plant Batch.
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/strain:
    put:
      tags:
      - PlantBatches
      summary: ''
      description: 'Changes the strain of plant batches at a specified Facility.</br><b>Permissions Required</b>: View Veg/Flower Plants • Manage Veg/Flower Plants Inventory'
      externalDocs:
        x-doc-key: put_plantbatches_v2_strain
      operationId: PutPlantBatchStrain
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to change the Strain of the Plant Batch.
        format: ''
      responses:
        200:
          description: Ok
  /plantbatches/v2/tag:
    put:
      tags:
      - PlantBatches
      summary: ''
      description: 'Replaces tags for plant batches at a specified Facility.</br><b>Permissions Required</b>: View Veg/Flower Plants • Manage Veg/Flower Plants Inventory'
      externalDocs:
        x-doc-key: put_plantbatches_v2_tag
      operationId: PutPlantBatchTag
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to replace Plant Batch tags.
        format: ''
      responses:
        200:
          description: Ok
...